Proje genelinde kapsamlı kod yeniden düzenleme ve performans iyileştirmeleri
- Ana uygulama başlatma ve rota yönetiminde köklü değişiklikler
- Kategori, konum ve abonelik sağlayıcılarını güncelle
- Kullanıcı kimlik doğrulama ve profil kurulum akışını iyileştir
- Sohbet, mesaj ve reklam modellerini yeniden düzenle
- Hata yakalama ve hata ayıklama mekanizmalarını geliştir
- Performans ve kod okunabilirliğini artır

+++ lib/features/home/widgets/filter_view.dart
@@ -47,7 +47,7 @@ class _FilterViewState extends State<FilterView> {
         // Kullanıcının konumu ile ilanın konumu arasındaki mesafeyi hesapla
         if (advert.location == null || _currentUser?.location == null) return false;
 
-        double distance = _currentUser!.location!.distanceTo(advert.location!);
+        int distance = _currentUser!.location!.distanceTo(advert.location!);
         return distance <= _distance!;
       }).toList();
     }
